requirements.txt
pandas
gspread
oauth2client
requests
mysql-connector-python
python-dotenv
